move benchmark tests

+++ test/gas/gasBenchmark.t.sol
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.13;
 
-import {PaymentEscrowBase} from "../../base/PaymentEscrowBase.sol";
-import {PaymentEscrow} from "../../../src/PaymentEscrow.sol";
+import {PaymentEscrowBase} from "../base/PaymentEscrowBase.sol";
+import {PaymentEscrow} from "../../src/PaymentEscrow.sol";
 
 contract GasBenchmarkBase is PaymentEscrowBase {
-    uint256 internal constant BENCHMARK_AMOUNT = 100e6;
+    uint120 internal constant BENCHMARK_AMOUNT = 100e6;
     uint16 internal constant BENCHMARK_FEE_BPS = 100; // 1%
 
     PaymentEscrow.PaymentInfo internal paymentInfo;
